pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "A product that analyzes human interactions (calls, chats, emails, sms etc.)."
authors = ["Oleksandr Besan <oleksandr.besan@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
pydantic = "^2.6.4"
python-dotenv = "^1.0.1"
uvicorn = "^0.28.0"
loguru = "^0.7.2"
pydantic-settings = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
